fix: skill select & skillpool select
