+++ console/src/locales/ru.json
@@ -290,7 +290,10 @@
     "validUrl": "Пожалуйста, введите действительный URL, начинающийся с http:// или https://",
     "importedSkill": "Импортирован навык: {{name}}",
     "batchOperation": "Пакетная операция",
-    "exitBatch": "Выйти из режима"
+    "exitBatch": "Выйти из режима",
+    "noSearchResults": "Навыки, соответствующие вашему запросу, не найдены",
+    "filterByTag": "Фильтр по тегу",
+    "noTags": "Нет доступных тегов"
   },
   "skillPool": {
     "description": "Общий локальный пул навыков для встроенных навыков и загрузок из рабочих пространств. Внешние импорты (ClawHub, LobeHub) — отдельная функция.",
